# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Connect API
  version: 1.0.0
  description: |
    Production REST API specification for **Social Connect**, an Instagram-inspired full-stack social media application built with the MERN stack.
    
    ### Key Architecture & Security Features:
    - **Dual Token Authentication**: Short-lived JWT Access Tokens (`Authorization: Bearer <token>`) combined with rotated, HttpOnly Refresh Tokens (`refreshToken` cookie).
    - **Session Invalidation & Replay Detection**: Refresh tokens are family-tracked with automatic revocation of all sessions upon reuse attack detection.
    - **Bilateral Interaction Protections**: Immediate blocking enforcement across feeds, user search, direct messaging, comments, and stories.
    - **Content Moderation State Machine**: Multi-stage reporting queue and non-destructive moderation statuses (`ACTIVE`, `HIDDEN`, `REMOVED`).
    - **Deterministic Cursor-Based Pagination**: Stable chronological feed pagination utilizing compound `(createdAt, _id)` database indexes without offset drift.
    - **Immutable Audit Logging**: Compliance-grade, non-repudiable audit logs tracking authentication, administrative interventions, and security anomalies.
servers:
  - url: http://localhost:5000/api
    description: Local Development Server
...
paths:
...
  /users/{username}/posts:
    get:
      operationId: getUserPosts
      tags: [Posts]
      summary: Get user profile posts grid
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/UsernamePathParam'
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/LimitQueryParam'
      responses:
        '200':
          description: Paginated posts by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
...
  /posts:
    post:
      operationId: createPost
      tags: [Posts]
      summary: Create new post with media
      security:
        - bearerAuth: []
      description: Publishes a new post with multipart/form-data media items (up to 10 images, JPEG/PNG/WebP magic byte validated).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                caption:
                  type: string
                  maxLength: 2200
                  example: 'Golden hour in Tokyo'
                location:
                  type: string
                  maxLength: 100
                  example: 'Tokyo, Japan'
                media:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Array of up to 10 authentic image files
      responses:
        '201':
          description: Post published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error or invalid file signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Account suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      operationId: getPosts
      tags: [Posts]
      summary: Public posts feed
      parameters:
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/LimitQueryParam'
      responses:
        '200':
          description: List of active public posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /posts/{id}:
    get:
      operationId: getPostById
      tags: [Posts]
      summary: Get post by ID
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/PostIdPathParam'
      responses:
        '200':
          description: Post details with author profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '404':
          description: Post not found or moderated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      operationId: updatePost
      tags: [Posts]
      summary: Edit own post
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                caption:
                  type: string
                  maxLength: 2200
                location:
                  type: string
                  maxLength: 100
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '403':
          description: IDOR protection - only post owner can edit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: deletePost
      tags: [Posts]
      summary: Delete own post
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostIdPathParam'
      responses:
        '200':
          description: Post deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '403':
          description: Only post owner can delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /posts/{postId}/save:
    post:
      operationId: savePost
      tags: [Saves]
      summary: Save a post to private bookmarks
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostIdParam'
      responses:
        '200':
          description: Post saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
    delete:
      operationId: unsavePost
      tags: [Saves]
      summary: Remove post from private bookmarks
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostIdParam'
      responses:
        '200':
          description: Post unsaved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /posts/{postId}/save-status:
    get:
      operationId: getSaveStatus
      tags: [Saves]
      summary: Check if post is bookmarked by current user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostIdParam'
      responses:
        '200':
          description: Save status returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /users/me/saved-posts:
    get:
      operationId: getSavedPosts
      tags: [Saves]
      summary: List user's saved posts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/LimitQueryParam'
      responses:
        '200':
          description: Paginated saved posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'

  # ─── COMMENTS & REPLIES ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Standard JWT access token passed in Authorization header
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: Secure HttpOnly refresh token cookie
    skipRateLimit:
      type: apiKey
      in: header
      name: x-skip-rate-limit
      description: Bypasses rate limiting in testing environments
  parameters:
    PageQueryParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-indexed page number
    LimitQueryParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
      description: Items per page
    UsernamePathParam:
      name: username
      in: path
      required: true
      schema:
        type: string
        minLength: 3
        maxLength: 30
      description: Target username (case-insensitive)
    PostIdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: 24-character hexadecimal MongoDB ObjectId of the post
    PostIdParam:
      name: postId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: 24-character hexadecimal MongoDB ObjectId of the post
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          nullable: true
    ErrorEnvelope:
      type: object
      required: [success, message, error]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Resource not found
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
...